refactor: Consolidate duplicate ComprehensiveSecurityTest files
Merge unique tests from tests/Feature/Security/ComprehensiveSecurityTest.php
into tests/Security/ComprehensiveSecurityTest.php and delete the duplicate.

Changes:
- Add test_sensitive_data_not_exposed to Security/ComprehensiveSecurityTest
- Delete duplicate tests/Feature/Security/ComprehensiveSecurityTest.php

The tests/Security/ folder is the dedicated security test folder with more
comprehensive tests including data providers. The Feature/Security version
had 80%+ overlap with the following coverage by dedicated test files:
- Token expiration: tests/Feature/Security/TokenExpirationTest.php
- Token abilities: tests/Feature/Security/ApiScopeEnforcementTest.php
- Concurrent sessions: tests/Feature/Security/ConcurrentSessionTest.php
- SQL injection: tests/Security/Penetration/SqlInjectionTest.php
- XSS: tests/Security/Penetration/XssTest.php
- CSRF: tests/Security/Penetration/CsrfTest.php
