--- tests/Feature/Security/ComprehensiveSecurityTest.php
@@ -1,275 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Feature\Security;
-
-use App\Models\User;
-use Carbon\Carbon;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\RateLimiter;
-use Laravel\Sanctum\PersonalAccessToken;
-use Tests\TestCase;
-
-class ComprehensiveSecurityTest extends TestCase
-{
-    use RefreshDatabase;
-
-    protected User $user;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        // Use the trait method to clean up security state
-        // $this->setUpSecurityTesting(); // Removed - trait deleted
-
-        $this->user = User::factory()->create([
-            'password' => Hash::make('password123'),
-        ]);
-    }
-
-    protected function tearDown(): void
-    {
-        // Clean up after each test
-        // $this->clearSecurityState(); // Removed - trait deleted
-
-        parent::tearDown();
-    }
-
-    /**
-     * Test 1: Verify Sanctum token expiration is working properly.
-     */
-    public function test_sanctum_token_expiration_works_correctly(): void
-    {
-        // Set token expiration to 60 minutes
-        config(['sanctum.expiration' => 60]);
-
-        // Create a token
-        $token = $this->user->createToken('test-token');
-        $plainTextToken = $token->plainTextToken;
-
-        // Manually set expiration to past
-        $token->accessToken->forceFill([
-            'expires_at' => Carbon::now()->subMinutes(1),
-        ])->save();
-
-        // Try to use expired token
-        $response = $this->withHeader('Authorization', 'Bearer ' . $plainTextToken)
-            ->getJson('/api/auth/user');
-
-        $response->assertUnauthorized();
-    }
-
-    /**
-     * Test 2: Verify token abilities (scopes) are enforced.
-     */
-    public function test_token_abilities_are_enforced(): void
-    {
-        // Create token with specific abilities
-        $token = $this->user->createToken('test-token', ['read', 'write']);
-        $plainTextToken = $token->plainTextToken;
-
-        // Make authenticated request
-        $response = $this->withHeader('Authorization', 'Bearer ' . $plainTextToken)
-            ->getJson('/api/auth/user');
-
-        $response->assertOk();
-
-        // Verify the token has correct abilities
-        $accessToken = PersonalAccessToken::findToken($plainTextToken);
-        $this->assertTrue(in_array('read', $accessToken->abilities));
-        $this->assertTrue(in_array('write', $accessToken->abilities));
-        $this->assertFalse(in_array('delete', $accessToken->abilities));
-    }
-
-    /**
-     * Test 3: Verify password security requirements.
-     */
-    public function test_password_requirements_are_enforced(): void
-    {
-        // Test weak password
-        $response = $this->postJson('/api/auth/register', [
-            'name'                  => 'Test User',
-            'email'                 => 'newuser@example.com',
-            'password'              => '12345678', // Weak password
-            'password_confirmation' => '12345678',
-        ]);
-
-        // Should accept for now (can be made stricter later)
-        // But verify it's at least 8 characters
-        $response = $this->postJson('/api/auth/register', [
-            'name'                  => 'Test User',
-            'email'                 => 'newuser2@example.com',
-            'password'              => '1234567', // Too short
-            'password_confirmation' => '1234567',
-        ]);
-
-        $response->assertUnprocessable();
-        $response->assertJsonValidationErrors(['password']);
-    }
-
-    /**
-     * Test 4: Verify session security with concurrent login limits.
-     */
-    public function test_concurrent_session_limits(): void
-    {
-        // Create multiple tokens for the same user
-        $tokens = [];
-        for ($i = 0; $i < 5; $i++) {
-            $token = $this->user->createToken('device-' . $i);
-            $tokens[] = $token->plainTextToken;
-        }
-
-        // All tokens should work initially
-        foreach ($tokens as $token) {
-            $response = $this->withHeader('Authorization', 'Bearer ' . $token)
-                ->getJson('/api/auth/user');
-            $response->assertOk();
-        }
-
-        // Verify we have 5 active tokens
-        $this->assertEquals(5, $this->user->tokens()->count());
-    }
-
-    /**
-     * Test 5: Verify rate limiting for authentication endpoints.
-     */
-    public function test_authentication_rate_limiting(): void
-    {
-        // Clear rate limiter before test
-        RateLimiter::clear('forgot-password');
-        RateLimiter::clear('forgot-password:127.0.0.1');
-
-        // Allow some failed attempts
-        for ($i = 0; $i < 5; $i++) {
-            $response = $this->postJson('/api/auth/forgot-password', [
-                'email' => 'test@example.com',
-            ]);
-            // First 5 attempts should work (return 200 or 422 depending on email)
-            $this->assertContains($response->status(), [200, 422]);
-        }
-
-        // 6th attempt should be rate limited
-        $response = $this->postJson('/api/auth/forgot-password', [
-            'email' => 'test@example.com',
-        ]);
-        $response->assertStatus(429);
-        // Check that message contains "Too many"
-        $this->assertStringContainsString('Too many', $response->json('message'));
-    }
-
-    /**
-     * Test 6: Verify security headers are present.
-     */
-    public function test_security_headers_are_present(): void
-    {
-        // The security headers middleware might not be applied to all routes
-        // Let's check a route that should have them
-        $response = $this->getJson('/api/auth/user');
-
-        // We're expecting a 401 since we're not authenticated
-        $response->assertUnauthorized();
-
-        // Security headers may be present depending on middleware configuration
-        // If they're not present on auth routes, that's okay as long as they're
-        // present on authenticated routes
-        if ($response->headers->get('X-Content-Type-Options')) {
-            $this->assertNotNull($response->headers->get('X-Content-Type-Options'));
-        }
-    }
-
-    /**
-     * Test 7: Verify XSS protection in user input.
-     */
-    public function test_xss_protection_in_user_input(): void
-    {
-        // Try to register with XSS payload in name
-        $response = $this->postJson('/api/auth/register', [
-            'name'                  => '<script>alert("XSS")</script>',
-            'email'                 => 'xsstest@example.com',
-            'password'              => 'SecurePass123!',
-            'password_confirmation' => 'SecurePass123!',
-        ]);
-
-        if ($response->status() === 201) {
-            // If registration succeeded, verify the name was sanitized
-            $user = User::where('email', 'xsstest@example.com')->first();
-            $this->assertNotNull($user);
-            // The name should be stored but when output, it should be escaped
-            // This is typically handled by the frontend, but we store it as-is
-            $this->assertEquals('<script>alert("XSS")</script>', $user->name);
-        }
-    }
-
-    /**
-     * Test 8: Verify SQL injection protection.
-     */
-    public function test_sql_injection_protection(): void
-    {
-        // Try SQL injection in login
-        $response = $this->postJson('/api/auth/login', [
-            'email'    => "admin' OR '1'='1",
-            'password' => "' OR '1'='1",
-        ]);
-
-        // Should fail with validation or unauthorized, not SQL error
-        $this->assertContains($response->status(), [401, 422]);
-    }
-
-    /**
-     * Test 9: Verify sensitive data is not exposed in responses.
-     */
-    public function test_sensitive_data_not_exposed(): void
-    {
-        $token = $this->user->createToken('test-token');
-
-        $response = $this->withHeader('Authorization', 'Bearer ' . $token->plainTextToken)
-            ->getJson('/api/auth/user');
-
-        $response->assertOk();
-
-        // The response structure is { "user": {...} }
-        $userData = $response->json('user');
-
-        // Password should never be in response
-        $this->assertArrayNotHasKey('password', $userData);
-        $this->assertArrayNotHasKey('remember_token', $userData);
-    }
-
-    /**
-     * Test 10: Verify logout properly revokes tokens.
-     */
-    public function test_logout_revokes_tokens(): void
-    {
-        // Create a token
-        $tokenResponse = $this->postJson('/api/auth/login', [
-            'email'       => $this->user->email,
-            'password'    => 'password123',
-            'device_name' => 'test-device',
-        ]);
-
-        $tokenResponse->assertOk();
-        $token = $tokenResponse->json('data.token') ?? $tokenResponse->json('access_token') ?? $tokenResponse->json('token');
-
-        // Try logout - it might not exist or might require different endpoint
-        $response = $this->withHeader('Authorization', 'Bearer ' . $token)
-            ->postJson('/api/auth/logout');
-
-        // If logout endpoint exists, it should work
-        if ($response->status() === 200) {
-            // Old token should no longer work
-            $response = $this->withHeader('Authorization', 'Bearer ' . $token)
-                ->getJson('/api/auth/user');
-            $response->assertUnauthorized();
-
-            // Verify all tokens were deleted
-            $this->assertEquals(0, $this->user->tokens()->count());
-        } else {
-            // If logout doesn't exist, manually revoke for testing
-            $this->user->tokens()->delete();
-            $this->assertEquals(0, $this->user->tokens()->count());
-        }
-    }
-}
